+++ R/get_labels.R
@@ -122,7 +122,7 @@
 #' # create vector
 #' x <- c(1, 2, 3, 2, 4, NA)
 #' # add less labels than values
-#' x <- set_labels(x, c("yes", "maybe", "no"), force.values = FALSE)
+#' x <- set_labels(x, labels = c("yes", "maybe", "no"), force.values = FALSE)
 #' # get labels for labelled values only
 #' get_labels(x)
 #' # get labels for all values
@@ -141,7 +141,10 @@
 #'
 #' # create vector with unused labels
 #' data(efc)
-#' set_labels(efc$e42dep) <- c("independent" = 1, "dependent" = 4, "not used" = 5)
+#' efc$e42dep <- set_labels(
+#'   efc$e42dep,
+#'   labels = c("independent" = 1, "dependent" = 4, "not used" = 5)
+#' )
 #' get_labels(efc$e42dep)
 #' get_labels(efc$e42dep, drop.unused = TRUE)
 #' get_labels(efc$e42dep, include.non.labelled = TRUE, drop.unused = TRUE)