+++ R/get_na.R
@@ -42,8 +42,8 @@
 #' get_na(x, as.tag = TRUE)
 #'
 #' # replace only the NA, which is tagged as NA(c)
-#' replace_na(x, 2, tagged.na = "c")
-#' get_na(replace_na(x, 2, tagged.na = "c"))
+#' replace_na(x, value = 2, tagged.na = "c")
+#' get_na(replace_na(x, value = 2, tagged.na = "c"))
 #'
 #' # data frame as input
 #' y <- labelled(c(2:3, 3:1, tagged_na("y"), 4:1),