+++ R/read_write.R
@@ -111,15 +111,15 @@ read_spss <- function(path, atomic.to.fac = FALSE, tag.na = FALSE) {
           max.range.end <- max(na.range[!is.infinite(na.range)], na.rm = T)
           # we start with range up to highest value
           if (any(na.range == Inf) && min.range.start <= max(x, na.rm = TRUE)) {
-            x <- set_na(x, sort(stats::na.omit(unique(x[x >= min.range.start]))), as.tag = TRUE)
+            x <- set_na(x, value = sort(stats::na.omit(unique(x[x >= min.range.start]))), as.tag = TRUE)
           }
           # we start with range up to highest value
           if (any(na.range == -Inf) && max.range.end >= min(x, na.rm = TRUE)) {
-            x <- set_na(x, sort(stats::na.omit(unique(x[x <= max.range.end]))), as.tag = TRUE)
+            x <- set_na(x, value = sort(stats::na.omit(unique(x[x <= max.range.end]))), as.tag = TRUE)
           }
           # here we have no infinite value range
           if (!any(is.infinite(na.range))) {
-            x <- set_na(x, sort(stats::na.omit(unique(c(
+            x <- set_na(x, value = sort(stats::na.omit(unique(c(
               na.range[!is.infinite(na.range)], x[x >= min.range.start & x <= max.range.end]
             )))), as.tag = TRUE)
           }
@@ -278,14 +278,19 @@ write_data <- function(x, path, type, use.tagged.na, enc.to.utf8, version) {
       # get value labels
       val.lab <- get_labels(x[[i]], attr.only = F, include.values = "n")
       # encode to UTF-8, if requested
-      if (!is.null(val.lab)) x[[i]] <- set_labels(x[[i]], enc2utf8(val.lab))
-      if (!is.null(var.lab)) x[[i]] <- set_label(x[[i]], enc2utf8(var.lab))
+      if (!is.null(val.lab)) x[[i]] <- set_labels(x[[i]], labels = enc2utf8(val.lab))
+      if (!is.null(var.lab)) x[[i]] <- set_label(x[[i]], lab = enc2utf8(var.lab))
     }
     # convert variable to labelled factor, so it can be saved
-    x[[i]] <- suppressWarnings(to_label(x[[i]], add.non.labelled = TRUE,
-                                        prefix = FALSE, drop.na = !use.tagged.na))
+    x[[i]] <-
+      suppressWarnings(to_label(
+        x[[i]],
+        add.non.labelled = TRUE,
+        prefix = FALSE,
+        drop.na = !use.tagged.na
+      ))
     # set back variable label
-    x[[i]] <- set_label(x[[i]], var.lab, "label")
+    x[[i]] <- set_label(x[[i]], lab = var.lab, attr.string = "label")
     # check column name
     end.point <- colnames(x)[i]
     # if it ends with a dot, add a char. dot is invalid last char for SPSS