+++ R/rec.R
@@ -28,6 +28,7 @@
 #'           column names in a data frame depends on the function call:
 #'           \itemize{
 #'             \item recoded variables (\code{rec()}) will be suffixed with \code{"_r"}
+#'             \item recoded variables (\code{recode_to()}) will be suffixed with \code{"_r0"}
 #'             \item dichotomized variables (\code{dicho()}) will be suffixed with \code{"_d"}
 #'             \item grouped variables (\code{split_var()}) will be suffixed with \code{"_g"}
 #'             \item grouped variables (\code{group_var()}) will be suffixed with \code{"_gr"}