+++ R/rec_pattern.R
@@ -32,7 +32,7 @@
 #' # recode carers age into groups of width 5
 #' x <- rec(efc$c160age, recodes = rp$pattern)
 #' # add value labels to new vector
-#' set_labels(x) <- rp$labels
+#' x <- set_labels(x, labels = rp$labels)
 #' # watch result
 #' frq(x)
 #'