+++ R/to_value.R
@@ -48,14 +48,14 @@
 #'
 #' # for numeric factor levels, value labels will be used, if present
 #' dummy1 <- factor(c("3", "4", "6"))
-#' set_labels(dummy1) <- c("first", "2nd", "3rd")
+#' dummy1 <- set_labels(dummy1, labels = c("first", "2nd", "3rd"))
 #' dummy1
 #' to_value(dummy1)
 #'
 #' # for non-numeric factor levels, these will be used.
 #' # value labels will be ignored
 #' dummy2 <- factor(c("D", "F", "H"))
-#' set_labels(dummy2) <- c("first", "2nd", "3rd")
+#' dummy2 <- set_labels(dummy2, labels = c("first", "2nd", "3rd"))
 #' dummy2
 #' to_value(dummy2)
 #'
@@ -148,6 +148,6 @@ to_value_helper <- function(x, start.at, keep.labels) {
     new_value <- as.numeric(as.character(x))
   }
   # check if we should attach former labels as value labels
-  if (keep.labels) new_value <- set_labels(new_value, labels, force.labels = T)
+  if (keep.labels) new_value <- set_labels(new_value, labels = labels, force.labels = T)
   return(new_value)
 }